requirements.txt
asgiref==3.8.1
Django==5.2.12
djangorestframework==3.15.2
djoser==2.2.3
django-cors-headers==4.6.0
sqlparse==0.5.13.11.1
attrs==22.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
iniconfig==2.0.0
mysql-connector-python==8.0.33
oauthlib==3.3.1
packaging==23.0
pluggy==1.0.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
pytest==7.2.2
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
cloudinary==1.44.2
django-cloudinary-storage==0.3.0
Pillow==12.2.0
